Merge from trunk
Converting to L3 was missing setting the booleans that keep track of whether a value has been set or not. Thanks to Chris Myers for spotting and reporting this. It only happens if you do not save your file immediately after converting to L3 - so might have been hidden for some time.