Changed CaseSummaryVector.java to CaseSummaryMap.java.
In order to allow for gaps in the case numbers (which we need to do for the next release of test cases), I had to change the implementation of how the list of test cases is handled.  Previously it used a Vector, with the index of an element in the vector being the case number, but the implementation assumed contiguous numbering.  Now it uses a TreeMap, with keys being the case numbers and values being the case data objects.