ok - so sometimes a reference is not in the actual Model referenced BUT in a submodel of that Model
In my current code which validates without flattening this is impossible to determine - as I do not which to actual instantiate submodels etc. So if the reference is not found in the ReferencedModel but the ReferencedModel has a comp plugin with submodels it will give it teh benefit of the doubt and not report an issue.

If there is an issue downstream in this case then validating the flttened model should find it. I'll put that in next.