So, the weird thing about a selector node is that its *first* child sometimes needs to have parentheses around it, but never any other children, which behave like function children in that they never have to have parentheses added, because they're delimited by commas.
In this change, we add the child to the call to the 'hasUnambiguousGrammar' so that the selector (and anything else like it) can distinguish between those two cases.  The actual function call itself involved calls to 'getNumChildren' and 'getChild(0)', so those functions were made public and created as virtual functions on ASTBase, respectively.

Then, the rest of the 'hasUnambiguousGrammar' was changed so that function parents *and* function children were both declared unambiguous, as was any situation where the *child* was priority '8', which used to be only functions, but now include vectors, matrices, and selectors, all of which have unambiguous-to-parents grammar, and never need parentheses.