added a representsBvar function to emulate the old isBvar function
it has a different name as the original isBvar function would only return true for a bvar element that was read in using the MathML reader. If you create an ASTNode of type lambda and add children to it the isBvar function would have returned false for all of these whilst the representsBvar function keeps track of whether they are bvars or not.