Last few leak fixes for Spatial:  the plugin objects weren't deleting the old objects when asked to create a new one, when the new one was a singleton child.  (This problem did not occur in objects in the sbml/ directory that also had singleton children--those 'createChild' functions properly deleted the old object.)
Note that 'delete NULL' is a legal no-op in C++; there is no need for our typical 'if (x != NULL) delete x' idiom.

There's still one leak that I see in SampledField::getUncompressedLength() as used in test test_SpatialExtension_read_compressed.  Unfortunately, it uses the compression library, which I'm unfamiliar with, so I'm not sure where the leak may be coming from.